move some code, make decorators work properly
