+++ backend/src/packages/chaiNNer_external/external_stable_diffusion/automatic1111/img2img.py
@@ -35,7 +35,6 @@ from .. import auto1111_group
 verify_api_connection()
 
 
-@cached
 @auto1111_group.register(
     schema_id="chainner:external_stable_diffusion:img2img",
     name="Image to Image",
@@ -102,6 +101,7 @@ verify_api_connection()
             channels=3,
         ),
     ],
+    decorators=[cached],
 )
 def img_to_image_node(
     image: np.ndarray,
