+++ backend/src/packages/chaiNNer_external/external_stable_diffusion/automatic1111/outpainting.py
@@ -44,7 +44,6 @@ class OutpaintingMethod(Enum):
     OUTPAINTING_MK2 = 1
 
 
-@cached
 @auto1111_group.register(
     schema_id="chainner:external_stable_diffusion:img2img_outpainting",
     name="Outpaint",
@@ -162,6 +161,7 @@ class OutpaintingMethod(Enum):
             channels=3,
         ),
     ],
+    decorators=[cached],
 )
 def img_to_img_outpainting_node(
     image: np.ndarray,
