+++ backend/src/packages/chaiNNer_external/external_stable_diffusion/automatic1111/txt2img.py
@@ -31,7 +31,6 @@ from .. import auto1111_group
 verify_api_connection()
 
 
-@cached
 @auto1111_group.register(
     "chainner:external_stable_diffusion:txt2img",
     name="Text to Image",
@@ -83,6 +82,7 @@ verify_api_connection()
             channels=3,
         ),
     ],
+    decorators=[cached],
 )
 def text_to_image(
     prompt: Optional[str],
