+++ backend/src/packages/chaiNNer_external/external_stable_diffusion/automatic1111/upscaling.py
@@ -128,8 +128,8 @@ UPSCALER_MODE_LABELS = {
             channels=3,
         )
     ],
+    decorators=[cached],
 )
-@cached
 def upscale_node(
     image: np.ndarray,
     mode: UpscalerMode,
